requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.19
pyftpdlib==1.5.5
pyOpenSSL==19.1.0
python-dotenv==0.11.0
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
